Run Travis tests only for a single tag

+++ .travis/test.sh
@@ -18,7 +18,7 @@ count_volumes() {
 }
 
 
-echo $'\n********************* Testing'
+echo $'\n*************** Testing'
 
 source .travis/tags
 
@@ -27,25 +27,29 @@ set -e
 # Will stop any configuration
 trap './t3 stop --rm; docker volume rm typo3-root typo3-data & >/dev/null' EXIT
 
-# Run tests only for the first tag since all images tagged in this build are identical
+# Run tests
 for T in $TAGS; do
-    # TYPO3 standalone
-    t3 run -t $T
+    # TYPO3 standalone (only for TYPO3 v9.5)
+    if [ "$TYPO3_VER" = '9.5' ]; then
+        echo $'\n*************** '"$TRAVIS_REPO_SLUG:$T standalone"
+        t3 run -t $T
 
-    test $(count_containers 'typo3') -eq 1
-    test $(count_containers 'typo3-db') -eq 0
-    test $(count_volumes 'typo3-root') -eq 1
-    test $(count_volumes 'typo3-data') -eq 0
+        test $(count_containers 'typo3') -eq 1
+        test $(count_containers 'typo3-db') -eq 0
+        test $(count_volumes 'typo3-root') -eq 1
+        test $(count_volumes 'typo3-data') -eq 0
 
-    t3 stop --rm
+        t3 stop --rm
 
-    test $(count_containers 'typo3(-db)?') -eq 0
-    test $(count_volumes 'typo3-root') -eq 1
+        test $(count_containers 'typo3(-db)?') -eq 0
+        test $(count_volumes 'typo3-root') -eq 1
 
-    docker volume rm typo3-root >/dev/null
+        docker volume rm typo3-root >/dev/null
+    fi
 
     # TYPO3 + MariaDB/PostgreSQL
     for DB_TYPE in mariadb postgresql; do
+        echo $'\n*************** '"$TRAVIS_REPO_SLUG:$T with $DB_TYPE"
         t3 run -d $DB_TYPE -t $T
 
         test $(count_containers 'typo3(-db)?') -eq 2
@@ -58,4 +62,7 @@ for T in $TAGS; do
 
         docker volume rm typo3-root typo3-data >/dev/null
     done
+
+    # Run tests only for the first tag since all images tagged in this build are identical
+    break
 done
